Refactor:환경변수 주입
