+++ src/main/resources/application.yml
@@ -22,9 +22,9 @@ spring:
           # 서비스 제공자 식별자
           naver:
             # 서비스 제공자측에 저희가 어떤 서비스인지 인증하기 위한 값
-            client-id: 3tVKSO15tNGbkeZJf8eE # ${CLIENT_ID}
-            client-secret: zHvANLwWHH # ${CLIENT_SECRET}
-            redirect-uri: http://localhost:8080/login/oauth2/code/naver # ${REDIRECT_URL} #Call Back url
+            client-id: ${CLIENT_ID}
+            client-secret: ${CLIENT_SECRET}
+            redirect-uri: ${REDIRECT_URL} #Call Back url
             authorization-grant-type: authorization_code # 어떤 방식으로 access token을 받을지 정의
             client-authentication-method: client_secret_post # Client Id, Client Secret를 요청의 어디에 포함할지 정의. Body
             client-name: Naver
@@ -36,7 +36,7 @@ server:
   port: 8080
 
 jwt:
-  secret: aaaabbbsdifqbvaesoioegwaaaabbbsdifqbvaesoioegwaaaabbbsdifqbvaesasdfqve
+  secret: ${JWT_SECRET}
 
 springdoc:
   swagger-ui:
